Refactor: clientSessionProvider utils 폴더로 이동
