+++ src/app/layout.tsx
@@ -3,7 +3,7 @@ import Header from '@/components/Header/Header';
 import '@/styles/globals.css';
 import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
-import ClientSessionProvider from '@/components/ClientSessionProvider/ClientSessionProvider';
+import ClientSessionProvider from '@/utils/clientSessionProvider';
 
 const inter = Inter({ subsets: ['latin'] });
 
@@ -19,14 +19,14 @@ export default function RootLayout({
   return (
     <html lang="ko">
       <body className={inter.className}>
-        <div id="modal" />
-        <div className="flex min-h-screen flex-col">
-          <Header />
-          <div className="flex-1">
-            <ClientSessionProvider>{children}</ClientSessionProvider>
+        <ClientSessionProvider>
+          <div id="modal" />
+          <div className="flex min-h-screen flex-col">
+            <Header />
+            <div className="flex-1">{children}</div>
+            <Footer />
           </div>
-          <Footer />
-        </div>
+        </ClientSessionProvider>
       </body>
     </html>
   );
